More faithful reproduction
Some things, such as the use of epsilon, had been cut from the original DIRECT algorithm that turned out to be extremely important on functions where the global minimum isn't extremely close to the origin. Also, reversed the 64-to-32-bit change done earlier due to, frankly, pointlessness when epsilon is used.